+++ edison-jobs/src/main/java/de/otto/edison/jobs/status/JobStatusDetailIndicator.java
@@ -15,6 +15,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import static de.otto.edison.status.domain.Status.ERROR;
 import static de.otto.edison.status.domain.Status.OK;
 import static de.otto.edison.status.domain.Status.WARNING;
 import static java.time.OffsetDateTime.now;
@@ -31,15 +32,18 @@ public class JobStatusDetailIndicator implements StatusDetailIndicator {
     private final String name;
     private final String jobType;
     private final Optional<Duration> maxAge;
+    private final Status jobErrorMapping;
 
     public JobStatusDetailIndicator(final JobRepository jobRepository,
                                     final String name,
                                     final String jobType,
-                                    final Optional<Duration> maxAge) {
+                                    final Optional<Duration> maxAge,
+                                    final Status jobErrorMapping) {
         this.jobRepository = jobRepository;
         this.name = name;
         this.jobType = jobType;
         this.maxAge = maxAge;
+        this.jobErrorMapping = jobErrorMapping;
     }
 
     @Override
@@ -49,7 +53,7 @@ public class JobStatusDetailIndicator implements StatusDetailIndicator {
             return jobs.isEmpty() ? statusDetailWhenNoJobAvailable() : toStatusDetail(jobs.get(0));
         } catch (final Exception e) {
             LOG.error("could not retrieve job status");
-            return StatusDetail.statusDetail(name, Status.ERROR, "could not retrieve job status");
+            return StatusDetail.statusDetail(name, ERROR, "could not retrieve job status");
         }
     }
 
@@ -57,15 +61,26 @@ public class JobStatusDetailIndicator implements StatusDetailIndicator {
         Status status;
         String message;
 
-        if (JobInfo.JobStatus.OK.equals(jobInfo.getStatus()) && jobTooOld(jobInfo)) {
-            status = WARNING;
-            message = JOB_TOO_OLD_MESSAGE + (maxAge.isPresent() ? maxAge.get() : "N/A");
-        } else if (JobInfo.JobStatus.OK.equals(jobInfo.getStatus())) {
-            status = OK;
-            message = SUCCESS_MESSAGE;
-        } else {
-            status = WARNING;
-            message = ERROR_MESSAGE;
+        switch(jobInfo.getStatus()) {
+            case OK:
+                if(jobTooOld(jobInfo)) {
+                    status = WARNING;
+                    message = JOB_TOO_OLD_MESSAGE + (maxAge.isPresent() ? maxAge.get() : "N/A");
+                } else {
+                    status = OK;
+                    message = SUCCESS_MESSAGE;
+                }
+                break;
+
+            case ERROR:
+                status = jobErrorMapping;
+                message = ERROR_MESSAGE;
+                break;
+
+            case DEAD:
+            default:
+                status = WARNING;
+                message = ERROR_MESSAGE;
         }
         return StatusDetail.statusDetail(name, status, message, runningDetailsFor(jobInfo));
     }