+++ edison-jobs/src/test/java/de/otto/edison/jobs/configuration/JobConfigurationTest.java
@@ -1,0 +1,12 @@
+package de.otto.edison.jobs.configuration;
+
+import org.testng.annotations.Test;
+
+public class JobConfigurationTest {
+
+
+    @Test(expectedExceptions = IllegalArgumentException.class)
+    public void shouldErrorOnWrongStatusMapping() {
+        new JobConfiguration().parseStatusMapping("wrong");
+    }
+}